Optimized Axis3D compositions.
